Laravel Policy Successfully done!

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use Illuminate\Support\Facades\Auth;
 
+use function PHPUnit\Framework\is_null;
+
 class PostController extends Controller
 {
     /**
@@ -14,14 +16,14 @@ class PostController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Post::class);
         $posts = Post::where('user_id', auth::user()->id)->get();
         return view('resources.post.index', ['posts' => $posts]);
-        
     }
 
     /**
      * Show the form for creating a new resource.
-     */
+     */ 
     public function create()
     {
         return view('resources.post.create');
@@ -46,6 +48,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.show', ['post' => $post]);
     }
 
@@ -54,6 +57,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('view', $post);
         return view('resources.post.edit', ['post' => $post]);
     }
 
@@ -62,6 +66,7 @@ class PostController extends Controller
      */
     public function update(UpdatePostRequest $request, Post $post)
     {
+        $this->authorize('view', $post);
         $post->update([
             'user_id' => auth::user()->id,
             'subject' => $request->subject,
@@ -76,6 +81,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('view', $post);
         $post->delete();
         return redirect()->route('post.index')->with('message', 'Post Successfully Deleted!');
     }
